Moved input to commons-ui-core
