+++ packages/commons-ui-core/src/Input/Input.snap.js
@@ -3,14 +3,14 @@
 exports[`<Input /> renders unchanged 1`] = `
 <div>
   <div
-    class="MuiFormControl-root MuiTextField-root css-1vqv0ht-MuiFormControl-root-MuiTextField-root"
+    class="MuiFormControl-root MuiTextField-root css-7skgcz-MuiFormControl-root-MuiTextField-root"
   >
     <div
-      class="MuiInputBase-root MuiOutlinedInput-root MuiInputBase-colorPrimary MuiInputBase-formControl css-1e9tdcl-MuiInputBase-root-MuiOutlinedInput-root"
+      class="MuiInputBase-root MuiOutlinedInput-root MuiInputBase-colorPrimary MuiInputBase-formControl css-9ddj71-MuiInputBase-root-MuiOutlinedInput-root"
     >
       <input
         aria-invalid="false"
-        class="MuiInputBase-input MuiOutlinedInput-input css-s9ndwd-MuiInputBase-input-MuiOutlinedInput-input"
+        class="MuiInputBase-input MuiOutlinedInput-input css-1t8l2tu-MuiInputBase-input-MuiOutlinedInput-input"
         id=":r0:"
         type="text"
         value=""
